Add a shared context to avoid having to repeat common stuff for scenarios
